Subgraph: Rename `appIds` file to `constants`
