+++ .prettierignore
@@ -4,4 +4,4 @@ example
 contracts
 manifest
 subgraph.template.yaml
-appIds.template.ts
+constants.template.ts
