+++ packages/subgraph/package.json
@@ -18,7 +18,7 @@
     "manifest:rinkeby": "NETWORK=rinkeby yarn manifest",
     "manifest:mainnet": "NETWORK=mainnet yarn manifest",
     "manifest:xdai": "NETWORK=xdai yarn manifest",
-    "clean": "node ../../node_modules/rimraf/bin.js abis build generated subgraph.yaml src/appIds.ts",
+    "clean": "node ../../node_modules/rimraf/bin.js abis build generated subgraph.yaml src/constants.ts",
     "lint": "eslint --ext .ts,.json,.md,.yaml,.graphql",
     "lint:fix": "eslint --ext ts,.json,.md,.yaml,.graphql --fix"
   },
