+++ packages/subgraph/src/constants.template.ts
